refactor unittests
